requirements.txt
asgiref==3.2.3
bootstrap4==0.1.0
certifi==2019.9.11
cffi==1.14.0
chardet==3.0.4
construct==2.5.3
cryptography==2.8
defusedxml==0.6.0
Django==3.0.3
django-ckeditor==5.8.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-froala-editor==3.1.0
django-haystack==2.8.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-redis==4.11.0
djangorestframework==3.11.0
elasticsearch==2.4.0
future==0.18.1
haystack==0.30
idna==2.8
jieba==0.39
jsonfield==3.1.0
lxml==4.5.0
mysqlclient==1.4.6
oauthlib==3.1.0
pefile==2019.4.18
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
python-ptrace==0.9.4
python3-openid==3.1.0
pytz==2019.2
redis==3.5.2
requests==2.22.0
requests-oauthlib==1.3.0
simplejson==3.17.0
six==1.12.0
sqlparse==0.3.0
swapper==1.1.2.post1
urllib3==1.22
